refactor(swiftpm): Move `parseLockFile` to the model file
The function is more closely related to the model. Apply the now
possible visibility scope reductions.
